Used NSCache instead of NSMutableDictionary Fixed to decode downloaded image data in a separate thread Added new delegate for tracking download progress Added enum SDWebImageManualSetImage for image preprocessing

+++ SDWebImage/SDWebImageDownloader.h
@@ -25,7 +25,7 @@ extern NSString *const SDWebImageDownloadStopNotification;
     NSMutableData *imageData;
     id userInfo;
     BOOL lowPriority;
-    NSUInteger expectedSize;
+    long long expectedSize;
     BOOL progressive;
     size_t width, height;
 }
